requirements.txt
django
djangorestframework
prettyconf==1.2.3
psycopg2-binary==2.7.4
